relocate components
1. register and login to auth folder
2. all widget to components/widgets
3. header and footer to components/common
4. home-page is moved from the "main" folder above
