+++ client/src/app/layouts/auth/login/login.component.scss
@@ -1,0 +1,57 @@
+.container-login {
+  position: relative;
+  &-close {
+    position: absolute;
+    top: -40px;
+    right: -20px;
+  }
+  .navigation {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    margin-top: 20px;
+    margin-bottom: 20px;
+    &__link {
+      margin: 0 40px;
+      color: #b1b1b1;
+      font-weight: 500;
+      font-size: 16px;
+      text-transform: uppercase;
+      transition: 0.4s;
+      cursor: pointer;
+    }
+    .navigation__link--active {
+      transition: 0.4s;
+      color: #000;
+      transition: 0.4s;
+      position: relative;
+      &:after {
+        content: "";
+        transition: 0.4s;
+        background-color: #349948;
+        border-radius: 5px;
+        position: absolute;
+        bottom: -10px;
+        left: -25%;
+        width: 150%;
+        height: 3px;
+      }
+    }
+  }
+  &__form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    &-field {
+      width: 75%;
+      margin-top: 20px;
+      font-size: 20px;
+      background-color: transparent;
+      outline: none;
+    }
+    &-btn {
+      margin-top: 20px;
+      width: 50%;
+    }
+  }
+}
