+++ client/src/app/layouts/auth/register/register.component.ts
