+++ client/src/app/layouts/components/common/footer/footer.component.scss
