+++ client/src/app/layouts/components/common/header/header.component.ts
