+++ client/src/app/layouts/components/widgets/widget-auto-portal/widget-auto-portal.component.ts
