+++ client/src/app/layouts/components/widgets/widget-section-id/widget-section-id.component.scss
