--- client/src/app/layouts/main/home/home.component.scss
@@ -1,90 +1,0 @@
-.home {
-  &__content {
-    margin: 0 auto;
-    max-width: 1600px;
-    display: flex;
-    padding: 0 15px;
-    @media (max-width: 1600px) {
-      padding: 0 calc(5px + (15 - 5) * ((100vw - 320px) / (1600 - 320)));
-    }
-    &-aside {
-      width: 25%;
-      padding-right: 0.5%;
-      border-right: 1px solid #e9e9e9;
-      .aside-list {
-        &-item {
-          margin-top: 10px;
-          padding: 0 10px;
-          font-size: 16px;
-          .aside-list-link {
-            display: flex;
-            align-items: center;
-            cursor: pointer;
-            &:hover {
-              .aside-list-link-text {
-                color: #3f51b5;
-                text-decoration: underline;
-              }
-            }
-            &-image {
-              margin-right: 10px;
-              width: 24px;
-              height: 24px;
-            }
-            &-text {
-              color: #000;
-              margin: 0;
-            }
-          }
-        }
-      }
-    }
-    &-main {
-      width: 75%;
-      padding-left: 0.5%;
-      .image__container {
-        max-width: 1200px;
-        width: 100%;
-        height: 350px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        img {
-          max-width: 100%;
-          max-height: 100%;
-        }
-      }
-      .swiper {
-        max-width: 1200px;
-        width: 100%;
-        height: 400px;
-      }
-      .main-history {
-        &-title {
-          color: #535353;
-        }
-      }
-    }
-  }
-}
-// Media START ==========================================================================================
-@media (max-width: 1000px) {
-  .home {
-    &__content {
-      &-aside {
-        display: none;
-      }
-      &-main {
-        width: 100%;
-        padding-left: 0;
-        .image__container {
-          height: auto;
-        }
-        .swiper {
-          height: auto;
-        }
-      }
-    }
-  }
-}
-// Media END ============================================================================================
